requirements.txt
numpy>=1.24.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
